Move to a StackView for detail items.
TableViews have an internal scroll view and that gives us a smaller area
to put the detail's views into.  A full-sized StackView allows us to
play with the full height content of the details better.  Here we keep
the search field above the site detail views as much as possible.