Core Data model version that avoids transformable properties warning.
"One or more models in this application are using transformable properties with transformer names that are either unset, or set to NSKeyedUnarchiveFromDataTransformerName"

Instead of opening in safe/read-only mode, migrate the DB to the
recommended way of describing NSData fields.  This should be
backwards-compatible.  This also guarantees that any older DB versions
will be automatically migrated to the latest version, so we don't need
to worry about missing fields such as url & touchID.