DataSource update deferring.
Data source has to be updated prior to each operation if not performed
in batch. Data source has to be updated at the end of the batch if it is.
This is especially important because the index paths on which to perform
operations are different when performing batch versus individual updates.