Fix constraint issue with temporary size.
A view controller's view starts out with a default temporary size which
can conflict with the controller's constraints if they require a view
that is bigger than the default application window size.

To dodge this issue, calculate the constraint size from the freshly
loaded view before appearing it, setting it as the view controller's
view frame, causing it (the correct minimum size) to be used for the
temporary size instead of the default application window's size.