Fix weak ServiceItems disappearing when dataSource updates.
If a new ServiceItem in the dataSource is equal to an old ServiceItem,
the service cell is not updated with the new item, but the old
ServiceItem was nil'ed since it is only weakly referenced by the cell.

Make cells strongly reference ServiceItems and ensure they are only
equal to new ServiceItems if they reference the same MPService instance,
forcing the cell to use only current MPService instances.