Enforce zero-PII policy.
[REMOVED]   Don't send unauthenticated identifiers to diagnostic services.
[ADDED]     User identifier based on master key.

It's important for Spectre that we do not hold any identifiers on you
that can be abused for identification purposes.

The salt-hashed full-name identifier is not compatible with our zero-PII
policy. Even though it is not possible for the identifier to be reversed
into its original full name, and even though it is not possible for
external parties to request information on a user by repeating the
full-name hash (due to the secret salt), it *is* possible for Spectre to
recreate the identifier based on a user's full name, and it *is*
possible for external parties to *compel* Spectre into doing so.

The same is true for keyID-based identifiers - even though the keyID is
not public knowledge, it could potentially be gleaned from confiscating
a user's hardware device and subsequently compelling Spectre into
extracting statistical information on a user.

The user identifier is now determined based on the user's master key,
hashed against a secret salt. Consequently, the only way to determine
the identifier is to know both the user's full name and their master
password. This information is only known by the user themselves. As a
result, external parties cannot compose the identifier and compel
Spectre to reveal information on an individual.