DataSource indexPath corrections.
[FIXED]     Don't reload inside performBatchUpdates.
[FIXED]     Correct insert and move index paths.

Base indexPaths off of performBatchUpdates theory:
https://developer.apple.com/videos/play/wwdc2018/225/?time=2016

Reload and move operations on the same element is not possible since
reloads break down into insert and delete, and delete is incompatible
with move. Instead, reload after updates if necessary. Due to the
callback, this requires synchronization.
https://stackoverflow.com/a/38625897/58803

Ensure it's impossible to have no tableView or collectionView and
impossible to have both.  It was never supported and would break
synchronization.